# apidoc/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CatCafeApi",
    "description": "猫カフェのポータルサイトと管理画面で使用するAPIを定義",
    "version": "0.1.0"
  },
...
  "servers": [
    {
      "url": "http://{username}:{port}/{basePath}",
      "description": "ローカル環境",
      "variables": {
        "username": {
          "default": "localhost"
        },
        "port": {
          "enum": [
            "8080"
          ],
          "default": "8080"
        },
        "basePath": {
          "default": ""
        }
      }
    }
  ],
  "paths": {
...
    "/{cats}/store": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        }
      ],
      "post": {
        "tags": [
          "store"
        ],
        "summary": "店舗登録API",
        "description": "店舗を1件登録する",
        "operationId": "storeCreate",
        "requestBody": {
          "description": "リクエストボディ",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/store/request.json#/definitions/store_create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/success_response/postResponse.json#/definitions/post_object"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/stores": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        }
      ],
      "get": {
        "tags": [
          "store"
        ],
        "summary": "店舗一覧取得API",
        "description": "店舗を取得する",
        "operationId": "storeSearch",
        "parameters": [
          {
            "name": "store_ids",
            "in": "query",
            "description": "店舗ID",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/store/response.json#/definitions/store_search_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/stores/{store_id}": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        },
        {
          "name": "store_id",
          "in": "path",
          "description": "店舗ID",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }
      ],
      "get": {
        "tags": [
          "store"
        ],
        "summary": "店舗詳細取得API",
        "description": "店舗詳細を1件取得する",
        "operationId": "storeFind",
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/store/response.json#/definitions/store_find_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          },
          "404": {
            "description": "Idが見つからない"
          }
        }
      },
      "patch": {
        "tags": [
          "store"
        ],
        "summary": "店舗更新API",
        "description": "店舗を1件更新する",
        "operationId": "storeUpdate",
        "requestBody": {
          "description": "リクエストボディ",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/store/request.json#/definitions/store_update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/success_response/patchResponse.json#/definitions/patch_object"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          },
          "404": {
            "description": "Idが見つからない"
          },
          "409": {
            "description": "排他失敗"
          }
        }
      },
      "delete": {
        "tags": [
          "store"
        ],
        "summary": "店舗削除API",
        "description": "店舗を1件論理削除する",
        "operationId": "storeDelete",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "description": "バージョンフィールド",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "正常系"
          },
          "404": {
            "description": "Idが見つからない"
          },
          "409": {
            "description": "排他失敗"
          }
        }
      }
    },
...
  }
}